requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
cachetools==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
distlib==0.3.8
docstring_parser==0.16
filelock==3.15.4
gcloud==0.18.3
google==3.0.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud==0.34.0
google-cloud-aiplatform==1.60.0
google-cloud-automl==2.13.4
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.4
google-cloud-storage==2.18.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
grpc-google-iam-v1==0.13.1
grpcio==1.65.1
grpcio-status==1.62.2
httplib2==0.22.0
identify==2.6.0
idna==3.7
nodeenv==1.9.1
numpy==2.0.1
oauth2client==4.1.3
packaging==24.1
platformdirs==4.2.2
pre-commit==3.8.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
shapely==2.0.5
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
